ui: moved IStopViewer one level upwards to avoid package cycle
