+++ de.setsoftware.reviewtool.core/src/de/setsoftware/reviewtool/config/IReviewConfigurable.java
@@ -2,8 +2,8 @@ package de.setsoftware.reviewtool.config;
 
 import de.setsoftware.reviewtool.model.IReviewPersistence;
 import de.setsoftware.reviewtool.model.changestructure.IChangeSource;
+import de.setsoftware.reviewtool.ui.IStopViewer;
 import de.setsoftware.reviewtool.ui.dialogs.EndReviewExtension;
-import de.setsoftware.reviewtool.ui.views.IStopViewer;
 
 /**
  * Interface to decouple the dynamically configurable review plugin from the
