+++ de.setsoftware.reviewtool.tests/src/de/setsoftware/reviewtool/config/ConfigurationInterpreterTest.java
@@ -28,8 +28,8 @@ import org.xml.sax.InputSource;
 import de.setsoftware.reviewtool.base.ReviewtoolException;
 import de.setsoftware.reviewtool.model.IReviewPersistence;
 import de.setsoftware.reviewtool.model.changestructure.IChangeSource;
+import de.setsoftware.reviewtool.ui.IStopViewer;
 import de.setsoftware.reviewtool.ui.dialogs.EndReviewExtension;
-import de.setsoftware.reviewtool.ui.views.IStopViewer;
 
 public class ConfigurationInterpreterTest {
